trim and substr

+++ libft/ft_putstr_fd.c
@@ -6,7 +6,7 @@
 /*   By: jlacaze- <jlacaze-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/06 23:58:56 by jlacaze-          #+#    #+#             */
-/*   Updated: 2024/11/11 20:36:59 by jlacaze-         ###   ########.fr       */
+/*   Updated: 2024/11/21 20:11:45 by jlacaze-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_putstr_fd(char *str, int fd)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i])
